backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
aiosqlite
beautifulsoup4
requests
google-generativeai
python-dotenv
pydantic
pydantic-settings
asyncpg